# Server/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Modern Library API",
    "version": "1.0.0",
    "description": "API documentation for learning, marketplace, AI, partner, and live classroom modules."
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Request validation failed" },
              "details": { "type": "object" }
            }
          },
          "requestId": { "type": "string", "example": "a8e29a68-cfca-41e2-97ac-906f7b5131f5" }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rotate refresh token and issue new access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New access and refresh token issued" },
          "401": { "description": "Refresh token invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout and revoke tokens",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Logged out" } }
      }
    },
...
  }
}